Separated build command from plugin boilerplate.
